Latest daily object and test class work.

+++ ECS.MemberManager.Core.EF.Domain/Payment.cs
@@ -7,8 +7,7 @@ namespace ECS.MemberManager.Core.EF.Domain
     {
         [Required]
         public Person Person { get; set; }
-        
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public DateTime PaymentDate { get; set; }
         public DateTime PaymentExpirationDate { get; set; }
         [Required]
